Bug 2002776: fix intermittent test failure due to missing tools image
This fixes an intermittent test failure that happens when `openshift/tools`
image (provided by the openshift/tools imagestream) used as the shell-image
in CreateExecPodOrFail helper is missing or yet to be populated in the
registry.

Before prometheus tests are run, it ensures that the imagestream is present
before running the actual test.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>