Fix should only include cluster daemonsets that have maxUnavailable for Hypershift
The test allows DaemonSets to ignore this rule if they have a
labelSelector for master nodes. In hypershift there are never master
nodes, hence these DSes don't have a selector for that.

Running these workloads as DS is wrong as that results in them getting
scheduled onto all nodes instead of three master nodes. Both of the
DaemonSets in question are from the network and currently it is unclear
how we want to do network in the future, so we just add a static
exception for now.

Ref https://issues.redhat.com/browse/HOSTEDCP-257