test/extended/util/openshift/clusterversionoperator: Drop waitForUpgradeable
Since we began polling in 4d1a25aee9 (test/e2e/upgrade/adminack: Poll
gates for duration of update, 2021-11-29, #26649), there's an
increased chance that this test is poking at acks while something else
is going on in Upgradeable in parallel.  For example [1]:

  Dec  1 21:01:44.722: FAIL: Error while waiting for Upgradeable to go true, err=timed out waiting for the condition Upgradeable: Status=False, Reason=KubeletMinorVersion_KubeletMinorVersionUnsupportedNextUpgrade, Message="Cluster operator kube-apiserver should not be upgraded between minor versions: KubeletMinorVersionUpgradeable: Kubelet minor versions on 6 nodes will not be supported in the next OpenShift minor version upgrade.".

With the incoming logic, this test-case would say "great, nothing in
Upgradeable complaining about any acks", without getting hung up on
the fact that Upgradeable is complaining about something else (in this
case, Kubelet skew, which will shortly resolve once the machine-config
gets around to updating nodes and catches them up with the
API-server).

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/release-openshift-origin-installer-e2e-aws-upgrade-4.7-to-4.8-to-4.9-to-4.10-ci/1466135844220833792#1:build-log.txt%3A866