Alter junit test suite names for jobs with multiple junit xmls.
It appears that on upgrade jobs where we invoke openshift-tests multiple
times (pre/post upgrade etc), the resulting two junit xml files are
merged by testgrid, and may contain the same test names. Testgrid then
mistakenly marks a test as a flake when in fact, one of them was a fatal
failure. This ultimately led to sippy showing massively successful test
pass rates when in fact we were failing almost all of the time.

This change attempts to address by using different junit test suite
names for different openshift-tests sub-commands.

In the case of the upgrade suite we're chasing problems in, we run:

openshift-tests run-upgrade all
openshift-tests run openshift/conformance/parallel

Hopefully this change will differentiate the tests sufficiently in
testgrid and sippy.