Bug 2026488: Skip Duplicate openshift-controller-manager Events
The full build suite includes several tests that force rollouts of
openshift-controller-manager. This causes duplicate
"reason/SuccessfulDelete" events to be detected, even though these
events are being triggered by separate rollouts. To fully address this
issue, openshift-controller-manager should be installed as a Deployment
rather than a DameonSet. This has better performance semantics with
respect to rollouts, which may benefit the install and upgrade
experience and reduce these "SuccessfulDelete" events.

This change provides a mechanism for known duplicate events to be
linked to a specific test suite, in addition to a specific platform or
topology. The known problem here is limited to the openshift/build test
suite, which can flake on this check and generally fails when the suite
is run on a cluster that enables tech preview features.

This skip should be removed when Bug 2034984 is fixed.

See also:
- https://bugzilla.redhat.com/show_bug.cgi?id=2026488
- https://bugzilla.redhat.com/show_bug.cgi?id=2034984