Use Local instead of Cluster for ext traf policy
With OVN on GCP there there is an LB disruption when
a node is rebooted. Each node in the cluster handles
LB requests when the policy is Cluster and there is
a bug that requests will still be given to the rebooting
node as they are LB'd to each node. When that happens
the request will hang and this shows up as a disruption.
the bug still needs to be resolved as this is not a
problem in other environments (openshift-sdn, other
clouds like aws)

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>