Bug 2026802: copy api-request upperbound for cluster-monitoring-operator
This patch addresses the bug by increasing the api-request upperbound of
cluster-monitoring-operator by copying the values from `master` for only
those platforms where the upperbound is lower than in `master`.
I.E. This only updates the upperbound of AWS, Azure, and GCP and not
BareMetal, VSphere and OpenStack since the ones in release-4.9 is higher
than the ones in master branch.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>