Add test for OS updates being staged within 5-10m.
Recent test failure investigation led to a possibility of slow disk
speeds doing OS content extraction. Add a test to ensure that we move
from OSUpdateStarted to OSUpdateStaged within 5-10 minutes, which
primarily just encompasses extraction of the content.

If any node is over 10 minutes we will fail the test, if some are over 5
we'll flake instead.