operators: extend test for exposed cluster services by optional path
This tests confirms that some cluster services are exposed externally.
This commit changes the test for the Prometheus route and adds an
optional path component for the checked route.
Background is that the prometheus route was changed to no longer allow
access to the Prometheus UI (https://issues.redhat.com/browse/MON-1961).
This was implemented by adding a path restriction to the route.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>