Fix HAProxy tests on FIPS properly
Change existing HAProxy tests to use a FIPS-compatible, 2048-bit RSA key.
Add a new test to verify that using a FIPS-incompatible, 1024-bit RSA key
for the router's default certificate fails on FIPS clusters and succeeds on
non-FIPS clusters.

The certificate used in this new test comes from the default certificate
that is currently built into the router image (which is why the tests were
failing on FIPS).  The certificate used for the already existing tests was
generated using the following Shell commands on a RHEL 8.4 system with
OpenSSL 1.1.1g:

    openssl req -x509 -newkey rsa:2048 -days 3650 -keyout ca.key -out ca.crt -nodes -subj '/C=US/ST=SC/L=Default City/O=Default Company Ltd/OU=Test CA/CN=www.exampleca.com/emailAddress=example@example.com'
    openssl req -newkey rsa:2048 -nodes -keyout tls.key -out router.csr -subj '/CN=www.example.com/ST=SC/C=US/emailAddress=example@example.com/O=Example/OU=Example'
    printf 'basicConstraints=CA:FALSE\n' > router.cnf
    openssl x509 -req -days 3650 -in router.csr -signkey tls.key -CA ca.crt -CAcreateserial -CAkey ca.key -out tls.crt -clrext -extfile router.cnf
    cat tls.{crt,key} > default_pub_keys.pem

These commands generate a certificate that has the same parameters as the
old one except for the serial number, signature algorithm, validity period,
and key size.

This commit fixes bug 2047913.

https://bugzilla.redhat.com/show_bug.cgi?id=2047913

* test/extended/router/certs.go: New file with the new tests.
* test/extended/testdata/router/router-common.yaml: Change route-1 and
route-2 from non-TLS routes to edge-terminated TLS routes with
insecureEdgeTerminationPolicy: Allow.
* test/extended/testdata/router/router-override-domains.yaml:
* test/extended/testdata/router/router-override.yaml:
* test/extended/testdata/router/router-scoped.yaml:
* test/extended/testdata/router/weighted-router.yaml: Add a parameter for
the default certificate with the default value being a certificate with a
FIPS-compatible, 2048-bit RSA key.
* test/extended/testdata/bindata.go:
* test/extended/util/annotate/generated/zz_generated.annotations.go:
Regenerate.