Reenable wait on worker deletion and increase serial test timeout
With the fix to remove the extra worker deletion (https://github.com/openshift/origin/pull/26774)
there have been occasional test failures when the time to delete the extra worker
has exceeded the 15 minute test time. This adds back in the wait after the extra worker
is deleted and increases the test time to allow the deletion to complete.