Attempt to fix lost events with improved watch resource version handling.
We have identified that e2e testing can lose events, they will not be
processed by the monitoring code and yet they exist when we view
events.json in must-gather.

As implemented we believe that the watch can be terminated and restarted
when apiservers are rolling. The current code will just do a new event
list and grab the latest resource version to start watching from. This
leaves a window in time in which events can be lost.

The new code tracks the latest rv we've seen on every event coming in,
and if we need to re-establish our watch we do so with the last
observed.