Increase scale test timeout to 30 minutes
When executed on a baremetal platform, the provisioning and deprovisioning time for an extra worker could take up to ~20 minutes on the current hardware used for the CI, with some slightly variations. For this reason, the current 20 minutes timeout in some cases is not enough. Increasing to 30 minutes should give enough room also for covering execution on slower hardware