MGMT-9440: Fix single node serial tests API crash.
It has been witnessed and reported in https://bugzilla.redhat.com/show_bug.cgi?id=2059581 that the KubeApi becomes unexpectedly unreachable for a short time after the completion of the  Serial API resiliance test. While we await the resolution of the BZ, I have introduced a 60 second wait at the end of the test to ensure that the tests after do not break.
It is our intention that this wait will be removed after the resolution of the BZ.