build-quota test: learn more about cgroupsv2
In the build-quota test script that we use to confirm that memory limits
have been correctly set, do more of the work to find the right cgroup
memory limit to read when we're in a cgroupsv2 setting.

The script already used some bash-specific syntax, so make that clear in
the shebang line.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>