Allow CurrentVersionNotVerified alert
Add the alert 'CurrentVersionNotVerified' as an allowed alert to tests
that are making sure any unexpected alerts are not firing.

The 'CurrentVersionNotVerified' alert is supposed to be firing when
the verification of the current version of a cluster has failed and
the version was forcefully applied to the cluster. Versions of the
clusters used in the CI may not be signed, and in these cases, it is
expected for the verification to fail.