MGMT-9440: Fix single node serial tests API crash
This is a follow on from https://github.com/openshift/origin/pull/26868
where we introduced a working to prevent single-node serial tests from
failing.

This patch rectfies that by removing the artificial delay and patching
the root cause, which is due to the way errors are processed when making
a call to isApiReady. The existing implementation will "bubble up"
errors from the httpClient.Get call to the caller of isApiReady.
However, at this stage, errors are undesirable, we simply want to ask
the question "is the API ready or not? -- for any reason"

So I have simplified the method to return a boolean response even if
errors are encountered.