Attempt multiple different token requests
The token request with client header authorization does not encode the
header client identifier payload with URLEscape like the other request
methods. The current oauth-server implements it so but this behavior has
changed in newer versions of the dependency. So now the test which makes
the token requests can multiple requests to fetch the token and as long
as one of them succeeds the test is considered as succeeded.