util: curl connect-to requires bracketed IPv6 address
curl `--connect-to` allows you to direct a request to a specific server,
bypassing DNS.  This is used in various router tests to push the request
through the haproxy instance, but in an IPv6 environment `--connect-to`
needs to be bracketed.

This makes the `.Through` helper accept either bracketed or unbracketed
IPv6 addresses.