tolerate some disruption while getting the list of machines from the server
we tolerate some disruption until https://bugzilla.redhat.com/show_bug.cgi?id=2082778
is fixed and rely on the monitor for reporting (p99).
this is okay since we observe disruption during the upgrade jobs too,
the only difference is that during the upgrade job we don’t access the API except from the monitor.