Revert "Revert "NE-626: Add DNS CI coverage for golang and glibc resolver libraries""
This reverts commit 4a1e80297759a22a4cae03c3deb6171c0fd382eb.

Modifications to the previous PR:

Base image for Go test was changed from golang in Docker Hub to registry.redhat.io/ubi8/go-toolset.
A test for Go 1.16 and 1.17 were added, 1.18 was removed as the go-toolset image does not yet support 1.18.
A check for baremetal environment was added to skip the test, as this environment is disconnected and does not have any suitable images on the local registry to use for different Go versions.