Add some debugging information for mysterious exit 2 errors
Some jobs, like these two [0][1], are showing up as a failure, but there
is no clear reason for the failure other than the test container was
exit(2).

It appears to be happening after the suite run, this adds some
additional debug gathering data to help figure out where this is coming
from.

[0] https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/periodic-ci-openshift-release-master-ci-4.11-upgrade-from-stable-4.10-e2e-aws-ovn-upgrade/1524255508167397376
[1] https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/periodic-ci-openshift-release-master-ci-4.11-upgrade-from-stable-4.10-e2e-aws-ovn-upgrade/1524255509832536064