Ignore tunbr NodeNetworkInterfaceFlapping alert
tunbr is an ephemeral bridge interface that can go up an down depending
on if the ephemeral network is created/deleted, we shouldnt care if that
happens. Same as we dont care if a veth interface goes up/down as that
is going to happen with CNI pod events.

Looking in to what it will take to have tunbr automatically ignored
in these alerts and not fire them. When that work is done we can
revert this commit.

https://issues.redhat.com/browse/SDN-3008

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>