etcd vertical e2e: EnsureMemberRemoved tolerates some failures
The etcd client factory used by the method establishes a stream connection to an etcd pod.
Since a connection like this is considered a long-running connection
and isn't subjected to gracefully termination
and the fact it runs after deleting a machine gives a reason to tolerate some failures while getting a list of members.

Note that unavailability of the APIs and the etcd cluster will be reported by the monitor tool.