Detect packet capture interface for EgressIP tests.
Detect and use the packet capture interface for SDN and OVNK. In the
case of SDN, this is the default route of the nodes. In the case of
OVNK, replicate logic found in the OVNK repo to find the main system
interface that belongs to bridge br-ex. Using the physical egress
interface for packet captures fixes a bug where the same interface would
be accounted for 2x.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>