Test unprivileged builds
Test builds where we set BUILD_PRIVILEGED=false in the environment to
ask the build controller to ask nodes to run builds with fewer
privileges.

Verify that builds run with lower privileges by looking for a diagnostic
message that the builder logs at log level 2 or higher when it notices
that it's running in a user namespace with ID mappings that aren't the
node-level default mappings.

Add unprivileged variations to multiple tests that ensure that we can
still pass secrets to the builder, and that it can read them properly.

Revise default-privileged cases that assumed the names of builds that
now share a BuildConfig with an unprivileged test, since the tests, and
their builds, can happen in unexpected sequence.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>