Report late requests verification: Fix for Hypershift
The late requests validation is based on interpreting the KAS audit
logs. The audit logs are fetched through `oc must-gather`. This works
fine for self-hosted OCP, but with an external controlplane the KAS runs
as pod in a distinct management cluster and thus it is not possible to
get its audit logs from the cluster under test.

This change fixes the test for Hypershift by introduction dedicated
audit log fetching codepaths for selfhosted/hypershift.