test/extended/security: re-fetch test pod to get Spec.NodeName
CreateExecPodOrFail() doesn't return the pod that was actually
created, it just returns the template used with the Create() call.
Which means the NodeName isn't populated.

The test tried to schedule another pod on the same node as the
test pod, but that obviously doesn't work if Spec.NodeName is
blank because CreateExecPodOrFail() doesn't return one.