Stop treating DNS lookup timeouts as disruption.
This is occuring often in the build clusters at least for the time
being, but we can identify that this specific message is a local DNS
failure, and thus should not be treated as disruption in the cluster
under test.

Change to detect this specific message and treat it as a warning level
interval, not an error level. Tests and the disruption json artifacts we
write both automatically filter to only error level intervals, so these
worked as is, although there is an additional filter added in this PR
just for safety.

Warning intervals will still be graphed on the intervals page.

A new test is added to watch for these DNS lookups and flake if found,
which will help us determine if the problem is still on-going.