Show DNS lookup sampler problems on the disruption intervals chart.
In #27238 we began treating DNS lookup errors as disruption warnings not
errors, as they indicate a problem in the CI cluster not the cluster
under test.

I confirmed warnings would show in the graph, but did not realize this
was based on the message. When the message later was changed for these
warnings, they no longer appeared.

Update the message logic to use a slightly safer "reason" my original PR
added, and include the new reason.