Update authenticated users RBAC e2e test
Problem: OLM has historically communicated to users which operators
are available in a given namespace by creating a copy of an operator
CSV in each namespace that it is scoped to. This has caused performance
issues on large clusters with many namespaces. OLM provided users with
the means to disable Copied CSVs for operators scoped to all namespaces,
but console was unable to communicate which operators were available
globally. Console will look for csvs in the openshift namespace to
identify which operators are available globally for authenticated users.

The authenticated user group needs read permissions to view the
CSVs in the openshift namespace. The extended e2e origin test suite has
a test that ensures that authenticated users do not have unanticipated
RBAC.

Solution: This commit updates the test so that it expects the authenticated
user group to have get, list, and watch permissions on CSVs in the openshift
namespace.