Test we don't see substantial Secret/ConfigMap growth during upgrade.
From a recent regression we looked into whether or not we could catch
what happened in a build cluster, in CI runs. I found that when this bug
was in play, Secret totals would grow by about 60% during upgrade
(roughly 600), and once the change was reverted this was not the case.

As such this test attempts to check for no more than a percentage of
growth for resources that are quite numerous. (i.e. we don't care if a
10 resource count grows to 20), but we do care if a 100 grows to
 800.

We are unsure if this will be feasible across all upgrade jobs, this
test is experimental and for now is just a flake until we're confident
it's assumptions are accurate.