Remove test 'clone repository using git:// protocol should clone using git:// if no proxy is configured'
This test was previously marked as skipped because github disabled the
git:// protocol. I assume the intention was to find another way to test
the use of git:// (maybe with some other platform?) but the bug is
closed and doesn't appear there's any intention to fix the test.

Despite the skip, this test still shows up in TRT's top failers list
because the BeforeEach/AfterEach stuff is still executed and sometimes
fail when other catastrophic things happened in the cluster.

This test is just creating noise, so I believe it should just be
removed.