test/extended/prometheus: Consider telemeterClient.enabled
I'd disabled Telemetry for the bulk of the CI fleet in
openshift/release@3c1da8eb20 (OTA-740:
ci-operator/step-registry/ipi/conf/telemetry: Disable Telemetry
(openshift/release#32153), 2022-09-13).  But that lead to many
failures for:

  Prometheus when installed on the cluster should report telemetry if a cloud.openshift.com token is present

This commit extends the checks for Telemetry enablement to include the
monitoring-specific ConfigMap, as well as the previously-checked
pull-secret token.

I'm copy/pasting a subset of the monitoring configuration structure
instead of vendoring the config, because we aren't vendoring the
cluster-monitoring-operator in origin today, bumping to keep such
vendoring up to date would be tedious, and the monitoring config API
is unlikely to shift this knob around within the structure.  It would
be nice if the monitoring config schema moved into opensihft/api or
somewhere else where it would be more clear that it was a
cluster-admin-facing API, with the usual stability commitments, but
it's not there today.