Add Kuryr exception to "pods should successfully create sandboxes" test
Kuryr can take long to create Pods as it needs to create ports for these
pods in Neutron. If a pod gets deleted while this is being processed, a
warning Event will be generated. This causes the `[sig-network] pods
should successfully create sandboxes by adding pod to network` to fail.

This commit adds this case as a flake.