Check for console capability before launching it's disruption monitoring.
Adds new utilities to check for enabled and known capabilities. Check
also assumes that if the cluster doesn't know about the console
capability, it predates 4.12 and thus we should assume to do monitoring
for console as it was default then.

The test should then see no ingress- backends for console, and not fail.