Fix rare race condition in ImageStream admission tests.
Details in TRT-577. Essentially we bump the limit from 0 to 1,
immediately try to create an ImageStream expecting it to pass, but get
rejected as the old limit of 0 appears to still be in effect.

There is no observed Status on LimitRange to wait for, so I think the
best we can do here is a sleep.