Drastically reduce the jobs we enforce disruption on.
Our operating assumption was that we were only enforcing disruption
using P99 values, thus would expect to only fail 1% of jobs, and if you
violate that test then you likely have regressed.

We have discovered that the vast majority of job platform combinations
do not have the 100 minimum runs required to get a P99, and thus we're
actually failing jobs on as low as P50's, which causes very erratic
failures and retests.

We are now limiting the data going into the historical
query_results.json in openshift/ci-tools repo. Here we now assume that
if we do not have disruption data for the platform under test, we flake
and report why. Previously this was a failure. The flake will allow us
to find instances where this is occurring if anyone needs to.