Add jUnitSuite and status to e2e-test interval locators.
Today there is no clear way to search for failed tests in the regex
search in spyglass or the html files it uses because the status of the
test is not in the locator. Searching the messages is non-trivial and
probably too slow as there can be many per row.

There are typically 3 intervals for each e2e test, start, finish, and
the end result with proper start/stop time. This change adds
status/Passed or Failed or Flakes to the latter.

In addition, I have added the jUnitSuite to the locator as well, this
will make it a little more clear what invocation of openshift-tests
we're running when we view intervals html. (today we sort of just know
instinctively by its ordering or what tests it runs) This will be useful
for a coming PR where I plan to extract the failed test names
from intervals.