Writeout test failures summary artifact.
This will be used to submit to a new sippy risk analysis API to
determine how risky the test failures in this job run are. As with junit
xmls, there can be multiple as we often invoke openshift-tests multiple
times in one job run.

Includes a change to make timestamp suffixes on junit artifacts more
consistent.