Add a test suite suffix to the alert test name
This seems unnecessary at first, but some tools like spyglass don't
properly indicate the testsuite name.  While the best fix is to address
all of those components, we have not proven good at doing that.  We can
make significant progress by providing unique names for the alert
failures to prevent false "flakes" and properly present failures on the
UI and on ci-search.