EgressIP: Run HTTP requests to netexec pods instead of using tcpdump
Significantly simplify EgressIP tests. In order to do so, we must
create a standalone CloudPrivateIPConfig with an extra IP and then
attach the IP address temporarily to the node's br-ex (or in the case
of SDN main interface). This will serve as our "external" host that
we can run EgressIP queries against.
We then run agnhost netexec inside that "external" host in a host
network pod. Finally, we run requests against the agnhost netexec pod
(endpoint /clientip). agnhost netexec will then return the IP address
of the client that's dialing in.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>