Fix missing disruption uploads.
Discovered that we are getting no data for image registry since Dec 7th
due to PR #27574 which shut off the entire monitor if we have less than
the required 100 runs to appear in the file of allowances to enforce. I
guess at this time, we did not have enough runs with image registry for
4.13. (which is unusual)

Skip the test, but let the monitor run.