Fix missing disruption data again.
Previous attempt did not work, as we were still bypassing
FrameworkEventIntervals() which is needed to add the test summary, which
eventually results in the AdditionalEvents json files in artifacts,
which are then merged into the overall events and disruption data files.
Without calling it, we still get no disruption data uploaded.

Move the nil check into the ExpectNoDisruption function, and allow
FrameworkEventIntervals to run as it should.