E2E networking tests: add retry to EgressFW
EgressFW tests are flaking on metal jobs.
Its suspected the tests are failing because
when we apply an EgressFirewall obj, it does
not mean OVN has applied this policy to the whole
cluster but that the obj is applied to the K API.
EgressFirewall does not expose a status or any
other mechanism to indicate when it is applied
throughout the cluster, therefore we must retry
until we get the result we expect.

Signed-off-by: Martin Kennelly <mkennell@redhat.com>