Handle missing alert data better
The "at or above info" tests are supposed to be using query_results.json
data to determine the thresholds at which they should fail or flake.
However I've discovered that the hardcoded list of these alert tests
does not match the hardcoded list of alerts used to query and populate
that results file.

Additionally, it looks like we're not properly handling the "no data
case", I suspect we're treating no data as 0s allowed, and thus failing
even though we have no historical basis to do so.

It appears all test authors assumed their alerts are automatically
picked up by the system which is not true today.

This change modifies such that we now skip the test and log a message if
we attempt to run a test against historical data we don't have. Test
will no longer fail.