EgressIP: Run HTTP requests to netexec container on external VM
Spawn a RHEL virtual machine on the MachineNetwork which serves as an
external target for EgressIP tests. Run the agnhost container inside the
VM with podman, then send requests from OCP to the VM.

While complex, this is the only reliable and realistic way to test the
EgressIP feature. The previous approach of sending UDP packets to a
non-existing target and analyzing the packets with tcpdump on the sender
machine turned out to be insufficient as this did not test allow-listing
of outgoing traffic on instance ports.

Signed-off-by: Andreas Karis <ak.karis@gmail.com>