Allow cluster daemonsets to use maxSurge
Change the "[sig-arch] Managed cluster should only include cluster
daemonsets that have maxUnavailable update of 10 or 33 percent" test to
allow daemonsets to specify maxUnavailable 0 with maxSurge 10%.

The intention of this test is to verify that cluster components have
parameters configured for rolling updates to ensure a reasonable balance
between rapid rollout and minimum availability during upgrades.  At the
time this test was written, the only way to accomplish that goal was to set
an appropriate maxUnavailable value.  Since then, the maxSurge parameter
has been added to the available parameters for daemonsets (see
<https://github.com/kubernetes/enhancements/tree/master/keps/sig-apps/1591-daemonset-surge>).
Now, a daemonset can ensure rapid rollout and minimum availability using
maxSurge.  However, when maxSurge is used, maxUnavailable must be zero.
Thus this commit changes the test to allow maxUnavailable to be zero if
maxSurge is set to 10%.

* test/extended/operators/daemon_set.go: Don't mark a daemonset as invalid
if it has maxSurge set to 10%.  Include the maxSurge value in the violation
message when neither maxUnavailable nor maxSurge is set appropriately.
* cmd/openshift-tests/minimal.go:
* pkg/test/ginkgo/testNames.txt:
* test/extended/testdata/bindata.go:
* test/extended/util/annotate/generated/zz_generated.annotations.go: Update
test name.