test/extended/cli/admin: Test 'oc adm release extract --file image-references ...'
Fixture generated with:

  $ oc adm release extract --file image-references quay.io/openshift-release-dev/ocp-release:4.13.0-rc.0-x86_64 >test/extended/testdata/cli/test-release-image-references.json

with the test/extended/testdata/cli location recommended by Maciej to
avoid the long, stuttery test/extended/testdata/cmd/test/cmd/testdata.

Compare this output from a different release image with the broken oc [1]:

  {
    "kind": "ImageStream",
    "apiVersion": "image.openshift.io/v1",
    ...
      "dockerImageRepository": ""
    }
  }Extracted release payload created at 2022-10-13T12:24:43Z

I'd initially suggested we include this test suite in the oc
repository [2], where the fixture could be versioned alongside the
code that generates it (making it easy to touch both the code and the
fixture in one pull request if there is a need to make adjustments).
But Arda and Maciej prefer this location in the origin repository, so
the test can live alongside existing oc testing.

[1]: https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/logs/periodic-ci-openshift-multiarch-master-nightly-4.12-ocp-e2e-sdn-serial-aws-arm64/1580538617384669184/artifacts/release/artifacts/release-images-arm64-latest
[2]: https://github.com/openshift/oc/pull/1317