e2e: node: testcase to exercise device manager recovery
When kubelet recovers pods, there's no guarantee about order of
recovering. This means a pod requesting devices can be recovered, and
can start, before the device plugin governing the device is started.

The end result is that containers can be admitted and start running without
the requested device actually available, causing runtime errors in the workload.
This is u/s issue
https://github.com/kubernetes/kubernetes/issues/109595

With this commit, we add a e2e test, focusing on the most critical SNO
scenario, which exercise the flow and verify the system behavior is
fixed and workload pods requesting devices are rejected (not admitted)
if they are recovered before the relevant device plugin.

This test want to force a node reboot like in `reboot command`
and targets SNO specifically.
Because of its high disruption potential, it must be opted-in
and doesn't run by default.
To enable the test, the test harness must supply a value for
the `DEVICE_RECOVERY_TARGET_NODE`, variable pointing to a node
in the cluster, SNO or worker node.

Signed-off-by: Francesco Romani <fromani@redhat.com>