machines: add a test which verifies number of node reboots
This adds a new test which checks `journalctl --list-boots` to ensure no unexpected reboots occurred on the nodes. It would run on SNO clusters only

Its similar to `should report ready nodes the entire duration of the test run` except that its applicable to SNO - when the only node reboots it won't be recorded as a change in prometheus metric.