UPSTREAM: <carry>: e2e/node/taints: Add toleration for DNS
In the "NoExecuteTaintManager" tests, configure DNS pods to tolerate the
taint that these tests add, so that the tests don't trigger a flood of
"TopologyAwareHintsDisabled" events by evicting DNS pods.

The DNS operator annotates the cluster DNS service to enable the
topology-aware hints feature.  However, this feature emits pathologically
repeating events when a DNS pod is evicted from a node.  These
pathologically repeating events trigger a test failure.  Thus in order to
allow the DNS operator to enable the topology-aware hints feature, it is
necessary to prevent tests from evicting DNS pods.

This commit fixes bug OCPBUGS-5943.

https://issues.redhat.com/browse/OCPBUGS-5943

* vendor/k8s.io/kubernetes/test/e2e/node/taints.go: Ensure DNS pods
tolerate the taint that the tests add, and remove the toleration once the
tests terminate.
* deps.diff: Regenerate.