[sig-kubevirt] Fix retrieval of virt-launcher pod of the guest node
Since the console-logger as been introduced, there is one more pod in the HCP namespace that contains the node name. We need to fix the filtering to get only the virt-launcher pods and then find the one that contains the node name amongst them, by the kubevirt.io=virt-launcher label.

Signed-off-by: Oren Cohen <ocohen@redhat.com>