Adjust interval builder before we continue porting
Moved to focus on IntervalBuilding, not ConditionBuilding, we hope to
eliminate Conditions in the future. (move the fields up into Interval)

Add a Display hint, coarse grained, where the generating code can
indicate if something is intended to be displayed in the UI. i.e.
DisruptionBegan DisruptionEnded no, Disruption intervals calculated from
those yes.

Experimenting with a source for each interval so we know where it came
from. This may get removed if we find it too cumbersome. Intent is to
help user see what they're looking at, and help the UI filtering have
something more concrete to filter on beyond message contents.