Fix ClusterOperator tests that can only report failures
We have discovered that if there are no clusteroperator intervals in the
e2e/conformance test portion, this test is omitted entirely which breaks
test pass rates in a variety of tooling, as well as the aggregator. In
this case if the operator does trigger an unexpected transition in the
conformance stage, a fail is recorded, and the other jobs in aggregation
report no result, resulting in < 6 results to aggregate on and a
failure.

Switch to always checking for, and including tests for, all operators,
regardless if they have recorded any intervals.