Skip invariant load balancer test for OpenStack
The disruption Service tests that were previously run and
skipped for OpenStack on upgrades jobs, are now run on various
jobs unrelated to upgrades. This can cause the job run to fail when
proxy installations are run with OpenStack given floating ip is not
supported in this scenario. This commit fixes the issue by continue to
skips those tests for OpenStack.