Add stlaz and p0lyn0mial to OWNERS
Both Standa and I have a good track record of contributing to this repository.
Standa has managed to contribute 100 PRs, and I've contributed 154.

Additionally, both the auth team and the API team contribute to this repository.
So having an approver from these teams will enable them to work more efficiently.