Do not fail etcd storage path test for GVRs that are not served.
Under the assumption that this test is of the behavior "apis are stored at the expected etcd paths,"
it can't make a positive or negative determination unless a given resource is served. Allowing it to
pass in that case relieves the test from needing to know the complete set of served GVRs.

Co-authored-by: Ben Luddy <bluddy@redhat.com>