Introduce OLM V1 Tests
This commit introduces a series of tests that ensure OLM V1 is behaving
as expected on tech preview clusters.

The first test case ensures that Catalogd controller successfully
reconciles a catalog CR and makes it's content available for install.

The second test ensures that the operator controller successfully
reconciles an operator CR and installs the expected operator using
RukPak.

Signed-off-by: Alexander Greene <greene.al1991@gmail.com>