Port cluster operator state intervals
These are sourced from the ClusterOperatorMonitor intervals.

Found a suspected problem with how reason and message were used in
these. We seemed to be using the condition Message as a Reason flag, and
ignoring the source condition Reason. This feels unintentional, so the
code is updated to now respect reason and treat message as HumanMessage.