Remove timelines command reprocessing intervals
This looks like something that may have been used during development of the new monitortest framework but possibly wasn't meant to be committed.

It does not seem like we would expect the monitortest command to reprocess intervals. If so, it seems broken and is lacking documentation as to what's going on. Using this locally you get a bunch of intervals at the start of the chart with no from, causing it to try to display a year, leaving all your actual intervals in a sliver on the right.